Splitting chat into its component classes
